4.21

+++ osa4/controllers/blogs.js
@@ -1,22 +1,30 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+/* const User = require('../models/user')
+const jwt = require('jsonwebtoken') */
+const {userExtractor} = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body
-
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const {title, author, url, likes} = request.body
+  const user = request.user
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
+    title: title,
+    author: author,
+    url: url,
+    likes: likes === undefined ? 0 : likes,
+    user: user._id
   })
+
   const savedBlog = await blog.save()
-  response.status(201).json(savedBlog)
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog.toJSON())
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -29,22 +37,29 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-  const body = request.body
+  const {title, author, url, likes} = request.body
 
   const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
+    title: title,
+    author: author,
+    url: url,
+    likes: likes
   }
 
   const update = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
   response.json(update)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog && blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({error: 'Delete failed'})
+  }
 })
 
 module.exports = blogsRouter
